+++ src/classes/features/PreventPause.ts
@@ -1,9 +1,5 @@
 import { ButtonAction, InputHook } from "isaac-typescript-definitions";
-import {
-  CallbackCustom,
-  ModCallbackCustom,
-  isRoomDangerous,
-} from "isaacscript-common";
+import { CallbackCustom, ModCallbackCustom, game } from "isaacscript-common";
 import { isPreventPauseEnabled } from "../../config";
 import { RandomizerModFeature } from "../RandomizerModFeature";
 import { isRandomizerEnabled } from "./achievementTracker/v";
@@ -18,7 +14,7 @@ export class PreventPause extends RandomizerModFeature {
     ButtonAction.PAUSE, // 12
   )
   inputActionPause(): boolean | undefined {
-    return isRoomDangerous() ? false : undefined;
+    return isRoomClear() ? false : undefined;
   }
 
   @CallbackCustom(
@@ -27,6 +23,11 @@ export class PreventPause extends RandomizerModFeature {
     ButtonAction.CONSOLE, // 28
   )
   inputActionConsole(): boolean | undefined {
-    return isRoomDangerous() ? false : undefined;
+    return isRoomClear() ? false : undefined;
   }
+}
+
+export function isRoomClear(): boolean {
+  const room = game.GetRoom();
+  return room.IsClear();
 }