+++ docs/history.md
@@ -33,7 +33,7 @@
   - Mausoleum: #10248, #20248
 - Fixed the bug where the "Recent Achievements" menu would sometimes be in the reverse order.
 - Fixed the bug where pills from champions would not have proper seeds. (Thanks Neonomi)
-- !Fixed the bug where some settings in the Dead Sea Scrolls menu would not save properly. (Thanks setJy)
+- Fixed the bug where some settings in the Dead Sea Scrolls menu would not save properly. (Thanks setJy)
 - Fixed the bug where alt-tabbing would let you illegally pause the game. (Thanks CorvusOnStream)
 - Fixed the bug where the Backasswards challenge would not work properly. (Thanks Andrew)